change capitalization of Kokkos to kokkos
If a user wants to use submodules like the AresPK documentation
suggests, there will be an issue with the capitalization of Kokkos in
CMakeLists.txt. the submodule creates a file external/kokkos, which is
not found with the current CMakeListst.txt. This is not a problem on
mac, since by default capitalization of paths does not matter, but on
linux it does matter. This fix should work on all machines since the
capitalization matches that of the submodule.